chore: grouped current version into treenode v1
