+++ packages/pluggableWidgets/tree-node-web/src/components/v1/Icons.tsx
