+++ packages/pluggableWidgets/tree-node-web/src/components/v1/TreeNode.tsx
@@ -2,7 +2,7 @@ import classNames from "classnames";
 import { ObjectItem, WebIcon } from "mendix";
 import { CSSProperties, ReactElement, ReactNode, useCallback, useContext } from "react";
 
-import { OpenNodeOnEnum, TreeNodeContainerProps } from "../../typings/TreeNodeProps";
+import { OpenNodeOnEnum, TreeNodeContainerProps } from "../../../typings/TreeNodeProps";
 
 import { useTreeNodeFocusChangeHandler } from "./hooks/TreeNodeAccessibility";
 import { useTreeNodeRef } from "./hooks/useTreeNodeRef";
