+++ packages/pluggableWidgets/tree-node-web/src/components/v1/hooks/useAnimatedHeight.tsx
