+++ packages/pluggableWidgets/tree-node-web/src/components/v2/TreeNode.tsx
@@ -1,0 +1,6 @@
+import { ReactElement } from "react";
+import { TreeNodeContainerProps } from "../../../typings/TreeNodeProps";
+
+export function TreeNodeV2(_props: TreeNodeContainerProps): ReactElement {
+    return <div />;
+}
