+++ packages/pluggableWidgets/tree-node-web/typings/TreeNodeProps.d.ts
@@ -4,7 +4,7 @@
  * @author Mendix Widgets Framework Team
  */
 import { ComponentType, CSSProperties, ReactNode } from "react";
-import { DynamicValue, ListValue, ListExpressionValue, ListWidgetValue, WebIcon } from "mendix";
+import { DynamicValue, ListValue, ListExpressionValue, ListReferenceValue, ListWidgetValue, WebIcon } from "mendix";
 
 export type HeaderTypeEnum = "text" | "custom";
 
@@ -19,6 +19,7 @@ export interface TreeNodeContainerProps {
     tabIndex?: number;
     advancedMode: boolean;
     datasource: ListValue;
+    parentAssociation?: ListReferenceValue;
     headerType: HeaderTypeEnum;
     openNodeOn: OpenNodeOnEnum;
     headerContent?: ListWidgetValue;
@@ -46,6 +47,7 @@ export interface TreeNodePreviewProps {
     translate: (text: string) => string;
     advancedMode: boolean;
     datasource: {} | { caption: string } | { type: string } | null;
+    parentAssociation: string;
     headerType: HeaderTypeEnum;
     openNodeOn: OpenNodeOnEnum;
     headerContent: { widgetCount: number; renderer: ComponentType<{ children: ReactNode; caption?: string }> };
